window-counter/requirements.txt
fastapi
uvicorn
streamlit
supervision
sahi
groundingdino-py
segment-anything
ultralytics
opencv-python
numpy
Pillow
requests
